refactor(eval): rename runner layers
